Method for converting all colors into their complements added to class Image.
On the color wheel, the complementary color of a given color is the one on the opposite side of the circle.
HSV color space is perfect for the task since it represents colors in a wheel.
Each color of the image is converted into HSV color space, hue is appropriately calculated and the color is converted again in RGB.