Method for image enhancement using global histogram equalization and range offset for brightness (improved histogram stretching method) implemented.
The proposed formulas for the method adjusted so that calculations run faster.
Problems with big numbers - solved by using operators <, > instead of calculating.
An output added.
Method gives exactly the same output as global HE (for couple of testing images) because the offset for adjusting brightness is pretty low. Turns out that global HE does the most of the job on these testing images.
I think (not really sure) that the method would give the better results if some more appropriate images are used (those on which global HE could not do much).